+++ src/Gateways/YSJKOPay.php
@@ -78,9 +78,9 @@ class YSJKOPay extends YSGatewayBase {
     protected function prepare_payment_data( $order, $pay_session ) {
         $data = parent::prepare_payment_data( $order, $pay_session );
 
-        // JKOPay doesn't support card binding
-        $data['confirm']['paymentBehavior'] = 'QuickPayment';
-        unset( $data['confirm']['paymentInstrument']['savePaymentInstrument'] );
+        // 街口支付使用一般付款，不需要卡片綁定
+        $data['confirm']['paymentBehavior'] = 'Regular';
+        unset( $data['confirm']['paymentInstrument'] );
 
         return $data;
     }